requirements.txt
feedparser==6.0.11
google-api-python-client==2.126.0
google-genai>=1.47.0
requests==2.31.0
python-dateutil==2.9.0
streamlit>=1.50.0
beautifulsoup4>=4.14.0
edge-tts>=7.2.0
nest-asyncio>=1.6.0
yt-dlp>=2025.1.1
youtube-transcript-api>=1.2.4
